update: change to a temporary directory before using wget to downloa the site

+++ infra/web.sh
@@ -1,4 +1,5 @@
 #!/bin/bash -xe
+set -e 
 sudo apt update
 sudo apt install wget unzip -y
 sudo apt install nginx -y
@@ -9,8 +10,13 @@ sudo systemctl enable nginx
 sudo systemctl status nginx
 cd /tmp 
 sudo wget https://www.tooplate.com/zip-templates/2137_barista_cafe.zip
-sudo unzip 2137_barista_cafe.zip -d /var/www/html
-sudo cp -r /var/www/html/2137_barista_cafe/* /var/www/html
-sudo nginx -s reload
-sudo systemctl restart nginx
+if [ -f "2137_barista_cafe.zip" ]; then
+    sudo unzip 2137_barista_cafe.zip -d /var/www/html
+    sudo cp -r /var/www/html/2137_barista_cafe/* /var/www/html
+    sudo nginx -s reload
+    sudo systemctl restart nginx
+else
+    echo "Something went wrong while downloading the file"
+fi
+
